add:出入金集計機能を追加 #27

+++ src/main/java/com/example/demo/dao/MoneyRecordDao.java
@@ -1,6 +1,8 @@
 package com.example.demo.dao;
 
 import java.io.Serializable;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 import org.springframework.stereotype.Repository;
@@ -10,5 +12,6 @@ public interface MoneyRecordDao <T> extends Serializable {
 	public List<T> getAll();
 	public List<T> findByUsername(String username);
 	public List<T> find(String fstr, String a);
-	public int sumMonthExpense(String fstr, String a);
+	public BigDecimal sumMonthExpense(String fstr, String a);
+
 }
